+++ src/ic-websocket-cdk/src/lib.rs
@@ -1,27 +1,25 @@
-use candid::{decode_one, encode_one, CandidType, Principal};
+use candid::{CandidType, Principal};
 use errors::WsError;
-#[cfg(not(test))]
-use ic_cdk::api::time;
-use ic_cdk::api::{caller, data_certificate, set_certified_data};
-use ic_cdk_timers::{clear_timer, set_timer, set_timer_interval, TimerId};
-use ic_certified_map::{labeled, labeled_hash, AsHashTree, Hash as ICHash, RbTree};
-use serde::{Deserialize, Serialize};
-use serde_cbor::Serializer;
-use sha2::{Digest, Sha256};
-use std::fmt;
-use std::panic;
-use std::rc::Rc;
-use std::time::Duration;
-use std::{
-    cell::RefCell,
-    collections::VecDeque,
-    collections::{HashMap, HashSet},
-    convert::AsRef,
-};
+
+use ic_cdk::api::caller;
+use serde::Deserialize;
 
 mod errors;
+mod state;
 mod tests;
+mod timers;
+mod types;
 mod utils;
+
+use state::*;
+use timers::*;
+use types::*;
+pub use types::{
+    CanisterWsCloseArguments, CanisterWsCloseResult, CanisterWsGetMessagesArguments,
+    CanisterWsGetMessagesResult, CanisterWsMessageArguments, CanisterWsMessageResult,
+    CanisterWsOpenArguments, CanisterWsOpenResult, CanisterWsSendResult, ClientPrincipal,
+    OnCloseCallbackArgs, OnMessageCallbackArgs, OnOpenCallbackArgs, WsHandlers, WsInitParams,
+};
 
 /// The label used when constructing the certification tree.
 const LABEL_WEBSOCKET: &[u8] = b"websocket";
@@ -39,996 +37,6 @@ const INITIAL_OUTGOING_MESSAGE_NONCE: u64 = 0;
 const INITIAL_CLIENT_SEQUENCE_NUM: u64 = 1;
 /// The initial sequence number for outgoing messages.
 const INITIAL_CANISTER_SEQUENCE_NUM: u64 = 0;
-
-pub type ClientPrincipal = Principal;
-#[derive(CandidType, Clone, Deserialize, Serialize, Eq, PartialEq, Debug, Hash)]
-pub(crate) struct ClientKey {
-    client_principal: ClientPrincipal,
-    client_nonce: u64,
-}
-
-impl ClientKey {
-    /// Creates a new instance of ClientKey.
-    fn new(client_principal: ClientPrincipal, client_nonce: u64) -> Self {
-        Self {
-            client_principal,
-            client_nonce,
-        }
-    }
-}
-
-impl fmt::Display for ClientKey {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}_{}", self.client_principal, self.client_nonce)
-    }
-}
-
-/// The result of [ws_open].
-pub type CanisterWsOpenResult = Result<(), String>;
-/// The result of [ws_close].
-pub type CanisterWsCloseResult = Result<(), String>;
-/// The result of [ws_message].
-pub type CanisterWsMessageResult = Result<(), String>;
-/// The result of [ws_get_messages].
-pub type CanisterWsGetMessagesResult = Result<CanisterOutputCertifiedMessages, String>;
-/// The result of [ws_send].
-pub type CanisterWsSendResult = Result<(), String>;
-
-/// The arguments for [ws_open].
-#[derive(CandidType, Clone, Deserialize, Serialize, Eq, PartialEq, Debug)]
-pub struct CanisterWsOpenArguments {
-    client_nonce: u64,
-    gateway_principal: GatewayPrincipal,
-}
-
-/// The arguments for [ws_close].
-#[derive(CandidType, Clone, Deserialize, Serialize, Eq, PartialEq, Debug)]
-pub struct CanisterWsCloseArguments {
-    client_key: ClientKey,
-}
-
-/// The arguments for [ws_message].
-#[derive(CandidType, Clone, Deserialize, Serialize, Eq, PartialEq, Debug)]
-pub struct CanisterWsMessageArguments {
-    msg: WebsocketMessage,
-}
-
-/// The arguments for [ws_get_messages].
-#[derive(CandidType, Clone, Deserialize, Serialize, Eq, PartialEq, Debug)]
-pub struct CanisterWsGetMessagesArguments {
-    nonce: u64,
-}
-
-/// Messages exchanged through the WebSocket.
-#[derive(CandidType, Clone, Debug, Deserialize, Serialize, Eq, PartialEq)]
-pub(crate) struct WebsocketMessage {
-    client_key: ClientKey, // The client that the gateway will forward the message to or that sent the message.
-    sequence_num: u64,     // Both ways, messages should arrive with sequence numbers 0, 1, 2...
-    timestamp: u64,        // Timestamp of when the message was made for the recipient to inspect.
-    is_service_message: bool, // Whether the message is a service message sent by the CDK to the client or vice versa.
-    #[serde(with = "serde_bytes")]
-    content: Vec<u8>, // Application message encoded in binary.
-}
-
-impl WebsocketMessage {
-    /// Serializes the message into a Vec<u8>, using CBOR.
-    fn cbor_serialize(&self) -> Result<Vec<u8>, String> {
-        let mut data = vec![];
-        let mut serializer = Serializer::new(&mut data);
-        serializer.self_describe().map_err(|e| e.to_string())?;
-        self.serialize(&mut serializer).map_err(|e| e.to_string())?;
-        Ok(data)
-    }
-}
-
-/// Element of the list of messages returned to the WS Gateway after polling.
-#[derive(CandidType, Clone, Debug, Deserialize, Serialize, Eq, PartialEq)]
-pub struct CanisterOutputMessage {
-    client_key: ClientKey, // The client that the gateway will forward the message to or that sent the message.
-    key: String,           // Key for certificate verification.
-    #[serde(with = "serde_bytes")]
-    content: Vec<u8>, // The message to be relayed, that contains the application message.
-}
-
-/// List of messages returned to the WS Gateway after polling.
-#[derive(CandidType, Clone, Debug, Deserialize, Serialize, Eq, PartialEq)]
-pub struct CanisterOutputCertifiedMessages {
-    messages: Vec<CanisterOutputMessage>, // List of messages.
-    #[serde(with = "serde_bytes")]
-    cert: Vec<u8>, // cert+tree constitute the certificate for all returned messages.
-    #[serde(with = "serde_bytes")]
-    tree: Vec<u8>, // cert+tree constitute the certificate for all returned messages.
-}
-
-type GatewayPrincipal = Principal;
-
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-/// Contains data about the registered WS Gateway.
-struct RegisteredGateway {
-    /// The queue of the messages that the gateway can poll.
-    messages_queue: VecDeque<CanisterOutputMessage>,
-    /// Keeps track of the nonce which:
-    /// - the WS Gateway uses to specify the first index of the certified messages to be returned when polling
-    /// - the client uses as part of the path in the Merkle tree in order to verify the certificate of the messages relayed by the WS Gateway
-    outgoing_message_nonce: u64,
-}
-
-impl RegisteredGateway {
-    /// Creates a new instance of RegisteredGateway.
-    fn new() -> Self {
-        Self {
-            messages_queue: VecDeque::new(),
-            outgoing_message_nonce: INITIAL_OUTGOING_MESSAGE_NONCE,
-        }
-    }
-
-    /// Resets the messages and nonce to the initial values.
-    fn reset(&mut self) {
-        self.messages_queue.clear();
-        self.outgoing_message_nonce = INITIAL_OUTGOING_MESSAGE_NONCE;
-    }
-
-    /// Increments the outgoing message nonce by 1.
-    fn increment_nonce(&mut self) {
-        self.outgoing_message_nonce += 1;
-    }
-}
-
-fn get_current_time() -> u64 {
-    #[cfg(test)]
-    {
-        0u64
-    }
-    #[cfg(not(test))]
-    {
-        time()
-    }
-}
-
-/// The metadata about a registered client.
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct RegisteredClient {
-    last_keep_alive_timestamp: u64,
-    gateway_principal: GatewayPrincipal,
-}
-
-impl RegisteredClient {
-    /// Creates a new instance of RegisteredClient.
-    fn new(gateway_principal: GatewayPrincipal) -> Self {
-        Self {
-            last_keep_alive_timestamp: get_current_time(),
-            gateway_principal,
-        }
-    }
-
-    /// Gets the last keep alive timestamp.
-    fn get_last_keep_alive_timestamp(&self) -> u64 {
-        self.last_keep_alive_timestamp
-    }
-
-    /// Set the last keep alive timestamp to the current time.
-    fn update_last_keep_alive_timestamp(&mut self) {
-        self.last_keep_alive_timestamp = get_current_time();
-    }
-}
-
-thread_local! {
-    /// Maps the client's key to the client metadata
-    /* flexible */ static REGISTERED_CLIENTS: Rc<RefCell<HashMap<ClientKey, RegisteredClient>>> = Rc::new(RefCell::new(HashMap::new()));
-    /// Maps the client's principal to the current client key
-    /* flexible */ static CURRENT_CLIENT_KEY_MAP: RefCell<HashMap<ClientPrincipal, ClientKey>> = RefCell::new(HashMap::new());
-    /// Keeps track of all the clients for which we're waiting for a keep alive message
-    /* flexible */ static CLIENTS_WAITING_FOR_KEEP_ALIVE: Rc<RefCell<HashSet<ClientKey>>> = Rc::new(RefCell::new(HashSet::new()));
-    /// Maps the client's key to the sequence number to use for the next outgoing message (to that client).
-    /* flexible */ static OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP: RefCell<HashMap<ClientKey, u64>> = RefCell::new(HashMap::new());
-    /// Maps the client's key to the expected sequence number of the next incoming message (from that client).
-    /* flexible */ static INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP: RefCell<HashMap<ClientKey, u64>> = RefCell::new(HashMap::new());
-    /// Keeps track of the Merkle tree used for certified queries
-    /* flexible */ static CERT_TREE: RefCell<RbTree<String, ICHash>> = RefCell::new(RbTree::new());
-    /// Keeps track of the principals of the WS Gateways that poll the canister
-    /* flexible */ static REGISTERED_GATEWAYS: RefCell<HashMap<GatewayPrincipal, RegisteredGateway>> = RefCell::new(HashMap::new());
-    /// The parameters passed in the CDK initialization
-    /* flexible */ static PARAMS: RefCell<WsInitParams> = RefCell::new(WsInitParams::default());
-    /// The acknowledgement active timer.
-    /* flexible */ static ACK_TIMER: Rc<RefCell<Option<TimerId>>> = Rc::new(RefCell::new(None));
-    /// The keep alive active timer.
-    /* flexible */ static KEEP_ALIVE_TIMER: Rc<RefCell<Option<TimerId>>> = Rc::new(RefCell::new(None));
-}
-
-/// Resets all RefCells to their initial state.
-fn reset_internal_state() {
-    let client_keys_to_remove: Vec<ClientKey> = REGISTERED_CLIENTS.with(|state| {
-        let map = state.borrow();
-        map.keys().cloned().collect()
-    });
-
-    // for each client, call the on_close handler before clearing the map
-    for client_key in client_keys_to_remove {
-        remove_client(&client_key);
-    }
-
-    // make sure all the maps are cleared
-    CURRENT_CLIENT_KEY_MAP.with(|map| {
-        map.borrow_mut().clear();
-    });
-    CLIENTS_WAITING_FOR_KEEP_ALIVE.with(|set| {
-        set.borrow_mut().clear();
-    });
-    OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut().clear();
-    });
-    INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut().clear();
-    });
-    CERT_TREE.with(|t| {
-        t.replace(RbTree::new());
-    });
-    REGISTERED_GATEWAYS.with(|map| {
-        for g in map.borrow_mut().values_mut() {
-            g.reset();
-        }
-    });
-}
-
-/// Resets the internal state of the IC WebSocket CDK.
-///
-/// **Note:** You should only call this function in tests.
-pub fn wipe() {
-    reset_internal_state();
-
-    custom_print!("Internal state has been wiped!");
-}
-
-fn register_gateway_if_not_present(gateway_principal: GatewayPrincipal) {
-    REGISTERED_GATEWAYS.with(|map| {
-        map.borrow_mut()
-            .entry(gateway_principal)
-            .or_insert_with(RegisteredGateway::new);
-    });
-}
-
-fn get_registered_gateway(
-    gateway_principal: &GatewayPrincipal,
-) -> Result<RegisteredGateway, String> {
-    REGISTERED_GATEWAYS.with(|map| {
-        map.borrow()
-            .get(gateway_principal)
-            .ok_or_else(|| WsError::GatewayNotRegistered { gateway_principal }.to_string())
-            .cloned()
-    })
-}
-
-fn is_registered_gateway(principal: &Principal) -> bool {
-    get_registered_gateway(principal).is_ok()
-}
-
-fn get_outgoing_message_nonce(gateway_principal: &GatewayPrincipal) -> Result<u64, String> {
-    get_registered_gateway(gateway_principal).map(|g| g.outgoing_message_nonce)
-}
-
-fn increment_outgoing_message_nonce(gateway_principal: &GatewayPrincipal) -> Result<(), String> {
-    REGISTERED_GATEWAYS.with(|map| {
-        map.borrow_mut()
-            .get_mut(gateway_principal)
-            .and_then(|g| {
-                g.increment_nonce();
-                Some(())
-            })
-            .ok_or_else(|| WsError::GatewayNotRegistered { gateway_principal }.to_string())
-    })
-}
-
-fn insert_client(client_key: ClientKey, new_client: RegisteredClient) {
-    CURRENT_CLIENT_KEY_MAP.with(|map| {
-        map.borrow_mut()
-            .insert(client_key.client_principal, client_key.clone());
-    });
-    REGISTERED_CLIENTS.with(|map| {
-        map.borrow_mut().insert(client_key, new_client);
-    });
-}
-
-fn get_registered_client(client_key: &ClientKey) -> Result<RegisteredClient, String> {
-    REGISTERED_CLIENTS.with(|map| {
-        map.borrow()
-            .get(client_key)
-            .ok_or_else(|| WsError::ClientKeyNotConnected { client_key }.to_string())
-            .cloned()
-    })
-}
-
-fn get_client_key_from_principal(client_principal: &ClientPrincipal) -> Result<ClientKey, String> {
-    CURRENT_CLIENT_KEY_MAP.with(|map| {
-        map.borrow()
-            .get(client_principal)
-            .cloned()
-            .ok_or_else(|| WsError::ClientPrincipalNotConnected { client_principal }.to_string())
-    })
-}
-
-fn check_registered_client_exists(client_key: &ClientKey) -> Result<(), String> {
-    get_registered_client(client_key).map(|_| ())
-}
-
-fn check_client_registered_to_gateway(
-    client_key: &ClientKey,
-    gateway_principal: &GatewayPrincipal,
-) -> Result<(), String> {
-    get_registered_client(client_key).and_then(|registered_client| {
-        registered_client
-            .gateway_principal
-            .eq(gateway_principal)
-            .then_some(())
-            .ok_or_else(|| {
-                WsError::ClientNotRegisteredToGateway {
-                    client_key,
-                    gateway_principal,
-                }
-                .to_string()
-            })
-    })
-}
-
-fn add_client_to_wait_for_keep_alive(client_key: &ClientKey) {
-    CLIENTS_WAITING_FOR_KEEP_ALIVE.with(|clients| {
-        clients.borrow_mut().insert(client_key.clone());
-    });
-}
-
-fn init_outgoing_message_to_client_num(client_key: ClientKey) {
-    OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut()
-            .insert(client_key, INITIAL_CANISTER_SEQUENCE_NUM);
-    });
-}
-
-fn get_outgoing_message_to_client_num(client_key: &ClientKey) -> Result<u64, String> {
-    OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP.with(|map| {
-        map.borrow().get(client_key).cloned().ok_or_else(|| {
-            WsError::OutgoingMessageToClientNumNotInitialized { client_key }.to_string()
-        })
-    })
-}
-
-fn increment_outgoing_message_to_client_num(client_key: &ClientKey) -> Result<(), String> {
-    OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut()
-            .get_mut(client_key)
-            .ok_or_else(|| {
-                WsError::OutgoingMessageToClientNumNotInitialized { client_key }.to_string()
-            })
-            .map(|n| *n += 1)
-    })
-}
-
-fn init_expected_incoming_message_from_client_num(client_key: ClientKey) {
-    INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut()
-            .insert(client_key, INITIAL_CLIENT_SEQUENCE_NUM);
-    });
-}
-
-fn get_expected_incoming_message_from_client_num(client_key: &ClientKey) -> Result<u64, String> {
-    INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP.with(|map| {
-        map.borrow().get(client_key).cloned().ok_or(
-            WsError::ExpectedIncomingMessageToClientNumNotInitialized { client_key }.to_string(),
-        )
-    })
-}
-
-fn increment_expected_incoming_message_from_client_num(
-    client_key: &ClientKey,
-) -> Result<(), String> {
-    INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut()
-            .get_mut(client_key)
-            .ok_or(
-                WsError::ExpectedIncomingMessageToClientNumNotInitialized { client_key }
-                    .to_string(),
-            )
-            .map(|n| *n += 1)
-    })
-}
-
-fn add_client(client_key: ClientKey, new_client: RegisteredClient) {
-    // insert the client in the map
-    insert_client(client_key.clone(), new_client);
-    // initialize incoming client's message sequence number to 1
-    init_expected_incoming_message_from_client_num(client_key.clone());
-    // initialize outgoing message sequence number to 0
-    init_outgoing_message_to_client_num(client_key);
-}
-
-fn remove_client(client_key: &ClientKey) {
-    CLIENTS_WAITING_FOR_KEEP_ALIVE.with(|set| {
-        set.borrow_mut().remove(client_key);
-    });
-    CURRENT_CLIENT_KEY_MAP.with(|map| {
-        map.borrow_mut().remove(&client_key.client_principal);
-    });
-    REGISTERED_CLIENTS.with(|map| {
-        map.borrow_mut().remove(client_key);
-    });
-    OUTGOING_MESSAGE_TO_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut().remove(client_key);
-    });
-    INCOMING_MESSAGE_FROM_CLIENT_NUM_MAP.with(|map| {
-        map.borrow_mut().remove(client_key);
-    });
-
-    let handlers = get_handlers_from_params();
-    handlers.call_on_close(OnCloseCallbackArgs {
-        client_principal: client_key.client_principal,
-    });
-}
-
-fn format_message_for_gateway_key(gateway_principal: &GatewayPrincipal, nonce: u64) -> String {
-    gateway_principal.to_string() + "_" + &format!("{:0>20}", nonce.to_string())
-}
-
-fn get_messages_for_gateway_range(
-    gateway_principal: &GatewayPrincipal,
-    nonce: u64,
-) -> (usize, usize) {
-    let max_number_of_returned_messages = get_params().max_number_of_returned_messages;
-
-    let messages_queue = get_registered_gateway(gateway_principal)
-        .unwrap() // the value exists because we just checked that the gateway is registered
-        .messages_queue;
-
-    let queue_len = messages_queue.len();
-
-    if nonce == 0 && queue_len > 0 {
-        // this is the case in which the poller on the gateway restarted
-        // the range to return is end=(last index) and start=max(end - max_number_of_returned_messages, 0)
-        let start_index = if queue_len > max_number_of_returned_messages {
-            queue_len - max_number_of_returned_messages
-        } else {
-            0
-        };
-
-        return (start_index, queue_len);
-    }
-
-    // smallest key used to determine the first message from the queue which has to be returned to the WS Gateway
-    let smallest_key = format_message_for_gateway_key(gateway_principal, nonce);
-    // partition the queue at the message which has the key with the nonce specified as argument to get_cert_messages
-    let start_index = messages_queue.partition_point(|x| x.key < smallest_key);
-    // message at index corresponding to end index is excluded
-    let mut end_index = queue_len;
-    if end_index - start_index > max_number_of_returned_messages {
-        end_index = start_index + max_number_of_returned_messages;
-    }
-    (start_index, end_index)
-}
-
-fn get_messages_for_gateway(
-    gateway_principal: &GatewayPrincipal,
-    start_index: usize,
-    end_index: usize,
-) -> Vec<CanisterOutputMessage> {
-    let messages_queue = get_registered_gateway(gateway_principal)
-        .unwrap() // the value exists because we just checked that the gateway is registered
-        .messages_queue;
-    let mut messages: Vec<CanisterOutputMessage> = Vec::with_capacity(end_index - start_index);
-    for index in start_index..end_index {
-        messages.push(
-            messages_queue
-                .get(index)
-                .unwrap() // the value exists because this function is called only after partitioning the queue
-                .clone(),
-        );
-    }
-    messages
-}
-
-/// Gets the messages in [MESSAGES_FOR_GATEWAYS] starting from the one with the specified nonce
-fn get_cert_messages(
-    gateway_principal: &GatewayPrincipal,
-    nonce: u64,
-) -> CanisterWsGetMessagesResult {
-    let (start_index, end_index) = get_messages_for_gateway_range(gateway_principal, nonce);
-    let messages = get_messages_for_gateway(gateway_principal, start_index, end_index);
-
-    if messages.is_empty() {
-        return get_cert_messages_empty();
-    }
-
-    let first_key = messages.first().unwrap().key.clone();
-    let last_key = messages.last().unwrap().key.clone();
-    let (cert, tree) = get_cert_for_range(&first_key, &last_key);
-
-    Ok(CanisterOutputCertifiedMessages {
-        messages,
-        cert,
-        tree,
-    })
-}
-
-fn get_cert_messages_empty() -> CanisterWsGetMessagesResult {
-    Ok(CanisterOutputCertifiedMessages {
-        messages: Vec::new(),
-        cert: Vec::new(),
-        tree: Vec::new(),
-    })
-}
-
-fn put_cert_for_message(key: String, value: &Vec<u8>) {
-    let root_hash = CERT_TREE.with(|tree| {
-        let mut tree = tree.borrow_mut();
-        tree.insert(key.clone(), Sha256::digest(value).into());
-        labeled_hash(LABEL_WEBSOCKET, &tree.root_hash())
-    });
-
-    set_certified_data(&root_hash);
-}
-
-fn push_message_in_gateway_queue(
-    gateway_principal: &GatewayPrincipal,
-    message: CanisterOutputMessage,
-) -> Result<(), String> {
-    REGISTERED_GATEWAYS.with(|map| {
-        // messages in the queue are inserted with contiguous and increasing nonces
-        // (from beginning to end of the queue) as ws_send is called sequentially, the nonce
-        // is incremented by one in each call, and the message is pushed at the end of the queue
-        map.borrow_mut()
-            .get_mut(gateway_principal)
-            .and_then(|g| {
-                g.messages_queue.push_back(message);
-                Some(())
-            })
-            .ok_or_else(|| WsError::GatewayNotRegistered { gateway_principal }.to_string())
-    })
-}
-
-fn get_cert_for_range(first: &String, last: &String) -> (Vec<u8>, Vec<u8>) {
-    CERT_TREE.with(|tree| {
-        let tree = tree.borrow();
-        let witness = tree.value_range(first.as_ref(), last.as_ref());
-        let tree = labeled(LABEL_WEBSOCKET, witness);
-
-        let mut data = vec![];
-        let mut serializer = Serializer::new(&mut data);
-        serializer.self_describe().unwrap();
-        tree.serialize(&mut serializer).unwrap();
-        (data_certificate().unwrap(), data)
-    })
-}
-
-fn put_ack_timer_id(timer_id: TimerId) {
-    ACK_TIMER.with(|timer| timer.borrow_mut().replace(timer_id));
-}
-
-fn reset_ack_timer() {
-    if let Some(t_id) = ACK_TIMER.with(Rc::clone).borrow_mut().take() {
-        clear_timer(t_id);
-    }
-}
-
-fn put_keep_alive_timer_id(timer_id: TimerId) {
-    KEEP_ALIVE_TIMER.with(|timer| timer.borrow_mut().replace(timer_id));
-}
-
-fn reset_keep_alive_timer() {
-    if let Some(t_id) = KEEP_ALIVE_TIMER.with(Rc::clone).borrow_mut().take() {
-        clear_timer(t_id);
-    }
-}
-
-fn reset_timers() {
-    reset_ack_timer();
-    reset_keep_alive_timer();
-}
-
-fn set_params(params: WsInitParams) {
-    PARAMS.with(|state| *state.borrow_mut() = params);
-}
-
-fn get_params() -> WsInitParams {
-    PARAMS.with(|state| state.borrow().clone())
-}
-
-fn get_handlers_from_params() -> WsHandlers {
-    get_params().get_handlers()
-}
-
-#[derive(CandidType, Debug, Deserialize, PartialEq, Eq)]
-pub(crate) struct CanisterOpenMessageContent {
-    client_key: ClientKey,
-}
-
-#[derive(CandidType, Debug, Deserialize, PartialEq, Eq)]
-pub(crate) struct CanisterAckMessageContent {
-    last_incoming_sequence_num: u64,
-}
-
-#[derive(CandidType, Debug, Deserialize, PartialEq, Eq)]
-pub(crate) struct ClientKeepAliveMessageContent {
-    last_incoming_sequence_num: u64,
-}
-
-/// A service message sent by the CDK to the client or vice versa.
-#[derive(CandidType, Debug, Deserialize, PartialEq, Eq)]
-pub(crate) enum WebsocketServiceMessageContent {
-    /// Message sent by the **canister** when a client opens a connection.
-    OpenMessage(CanisterOpenMessageContent),
-    /// Message sent _periodically_ by the **canister** to the client to acknowledge the messages received.
-    AckMessage(CanisterAckMessageContent),
-    /// Message sent by the **client** in response to an acknowledgement message from the canister.
-    KeepAliveMessage(ClientKeepAliveMessageContent),
-}
-
-impl WebsocketServiceMessageContent {
-    fn from_candid_bytes(bytes: &[u8]) -> Result<Self, String> {
-        decode_one(&bytes).map_err(|err| WsError::DecodeServiceMessageContent { err }.to_string())
-    }
-}
-
-fn send_service_message_to_client(
-    client_key: &ClientKey,
-    message: &WebsocketServiceMessageContent,
-) -> Result<(), String> {
-    let message_bytes = encode_one(&message).unwrap();
-    _ws_send(client_key, message_bytes, true)
-}
-
-/// Start an interval to send an acknowledgement messages to the clients.
-///
-/// The interval callback is [send_ack_to_clients_timer_callback]. After the callback is executed,
-/// a timer is scheduled to check if the registered clients have sent a keep alive message.
-fn schedule_send_ack_to_clients() {
-    let ack_interval_ms = get_params().send_ack_interval_ms;
-    let timer_id = set_timer_interval(Duration::from_millis(ack_interval_ms), move || {
-        send_ack_to_clients_timer_callback();
-
-        schedule_check_keep_alive();
-    });
-
-    put_ack_timer_id(timer_id);
-}
-
-/// Schedules a timer to check if the clients (only those to which an ack message was sent) have sent a keep alive message
-/// after receiving an acknowledgement message.
-///
-/// The timer callback is [check_keep_alive_timer_callback].
-fn schedule_check_keep_alive() {
-    let keep_alive_timeout_ms = get_params().keep_alive_timeout_ms;
-    let timer_id = set_timer(Duration::from_millis(keep_alive_timeout_ms), move || {
-        check_keep_alive_timer_callback(keep_alive_timeout_ms);
-    });
-
-    put_keep_alive_timer_id(timer_id);
-}
-
-/// Sends an acknowledgement message to the client.
-/// The message contains the current incoming message sequence number for that client,
-/// so that the client knows that all the messages it sent have been received by the canister.
-fn send_ack_to_clients_timer_callback() {
-    for client_key in REGISTERED_CLIENTS.with(Rc::clone).borrow().keys() {
-        // ignore the error, which shouldn't happen since the client is registered and the sequence number is initialized
-        match get_expected_incoming_message_from_client_num(client_key) {
-            Ok(expected_incoming_sequence_num) => {
-                let ack_message = CanisterAckMessageContent {
-                    // the expected sequence number is 1 more because it's incremented when a message is received
-                    last_incoming_sequence_num: expected_incoming_sequence_num - 1,
-                };
-                let message = WebsocketServiceMessageContent::AckMessage(ack_message);
-                if let Err(e) = send_service_message_to_client(client_key, &message) {
-                    // TODO: decide what to do when sending the message fails
-
-                    custom_print!(
-                        "[ack-to-clients-timer-cb]: Error sending ack message to client {}: {:?}",
-                        client_key,
-                        e
-                    );
-                } else {
-                    add_client_to_wait_for_keep_alive(client_key);
-                }
-            },
-            Err(e) => {
-                // TODO: decide what to do when getting the expected incoming sequence number fails (shouldn't happen)
-                custom_print!(
-                    "[ack-to-clients-timer-cb]: Error getting expected incoming sequence number for client {}: {:?}",
-                    client_key,
-                    e,
-                );
-            },
-        }
-    }
-
-    custom_print!("[ack-to-clients-timer-cb]: Sent ack messages to all clients");
-}
-
-/// Checks if the clients for which we are waiting for keep alive have sent a keep alive message.
-/// If a client has not sent a keep alive message, it is removed from the connected clients.
-fn check_keep_alive_timer_callback(keep_alive_timeout_ms: u64) {
-    let client_keys_to_remove: Vec<ClientKey> = CLIENTS_WAITING_FOR_KEEP_ALIVE
-        .with(Rc::clone)
-        .borrow()
-        .iter()
-        .filter_map(|client_key| {
-            // get the last keep alive timestamp for the client and check if it has exceeded the timeout
-            if let Some(client_metadata) =
-                REGISTERED_CLIENTS.with(Rc::clone).borrow().get(client_key)
-            {
-                let last_keep_alive = client_metadata.get_last_keep_alive_timestamp();
-                if get_current_time() - last_keep_alive > (keep_alive_timeout_ms * 1_000_000) {
-                    Some(client_key.to_owned())
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    for client_key in client_keys_to_remove {
-        remove_client(&client_key);
-
-        custom_print!(
-            "[check-keep-alive-timer-cb]: Client {} has not sent a keep alive message in the last {} ms and has been removed",
-            client_key,
-            keep_alive_timeout_ms
-        );
-    }
-
-    custom_print!("[check-keep-alive-timer-cb]: Checked keep alive messages for all clients");
-}
-
-fn handle_keep_alive_client_message(
-    client_key: &ClientKey,
-    _keep_alive_message: ClientKeepAliveMessageContent,
-) {
-    // TODO: delete messages from the queue that have been acknowledged by the client
-
-    // update the last keep alive timestamp for the client
-    if let Some(client_metadata) = REGISTERED_CLIENTS
-        .with(Rc::clone)
-        .borrow_mut()
-        .get_mut(client_key)
-    {
-        client_metadata.update_last_keep_alive_timestamp();
-    }
-}
-
-/// Internal function used to put the messages in the outgoing messages queue and certify them.
-fn _ws_send(
-    client_key: &ClientKey,
-    msg_bytes: Vec<u8>,
-    is_service_message: bool,
-) -> CanisterWsSendResult {
-    // get the registered client if it exists
-    let registered_client = get_registered_client(client_key)?;
-
-    // the nonce in key is used by the WS Gateway to determine the message to start in the polling iteration
-    // the key is also passed to the client in order to validate the body of the certified message
-    let outgoing_message_nonce = get_outgoing_message_nonce(&registered_client.gateway_principal)?;
-    let message_key = format_message_for_gateway_key(
-        &registered_client.gateway_principal,
-        outgoing_message_nonce,
-    );
-
-    // increment the nonce for the next message
-    increment_outgoing_message_nonce(&registered_client.gateway_principal)?;
-
-    // increment the sequence number for the next message to the client
-    increment_outgoing_message_to_client_num(client_key)?;
-
-    let websocket_message = WebsocketMessage {
-        client_key: client_key.clone(),
-        sequence_num: get_outgoing_message_to_client_num(client_key)?,
-        timestamp: get_current_time(),
-        is_service_message,
-        content: msg_bytes,
-    };
-
-    // CBOR serialize message of type WebsocketMessage
-    let message_content = websocket_message.cbor_serialize()?;
-
-    // certify data
-    put_cert_for_message(message_key.clone(), &message_content);
-
-    push_message_in_gateway_queue(
-        &registered_client.gateway_principal,
-        CanisterOutputMessage {
-            client_key: client_key.clone(),
-            content: message_content,
-            key: message_key,
-        },
-    )
-}
-
-fn handle_received_service_message(
-    client_key: &ClientKey,
-    content: &[u8],
-) -> CanisterWsMessageResult {
-    let decoded = WebsocketServiceMessageContent::from_candid_bytes(content)?;
-    match decoded {
-        WebsocketServiceMessageContent::OpenMessage(_)
-        | WebsocketServiceMessageContent::AckMessage(_) => {
-            WsError::InvalidServiceMessage.to_string_result()
-        },
-        WebsocketServiceMessageContent::KeepAliveMessage(keep_alive_message) => {
-            handle_keep_alive_client_message(client_key, keep_alive_message);
-            Ok(())
-        },
-    }
-}
-
-/// Arguments passed to the `on_open` handler.
-pub struct OnOpenCallbackArgs {
-    pub client_principal: ClientPrincipal,
-}
-/// Handler initialized by the canister
-/// and triggered by the CDK once the IC WebSocket connection is established.
-type OnOpenCallback = fn(OnOpenCallbackArgs);
-
-/// Arguments passed to the `on_message` handler.
-/// The `message` argument is the message received from the client, serialized in Candid.
-/// To deserialize the message, use [candid::decode_one].
-///
-/// # Example
-/// This example is the deserialize equivalent of the [ws_send's example](fn.ws_send.html#example) serialize one.
-/// ```rust
-/// use candid::{decode_one, CandidType};
-/// use ic_websocket_cdk::OnMessageCallbackArgs;
-/// use serde::Deserialize;
-///
-/// #[derive(CandidType, Deserialize)]
-/// struct MyMessage {
-///     some_field: String,
-/// }
-///
-/// fn on_message(args: OnMessageCallbackArgs) {
-///     let received_message: MyMessage = decode_one(&args.message).unwrap();
-///
-///     println!("Received message: some_field: {:?}", received_message.some_field);
-/// }
-/// ```
-pub struct OnMessageCallbackArgs {
-    /// The principal of the client sending the message to the canister.
-    pub client_principal: ClientPrincipal,
-    /// The message received from the client, serialized in Candid. See [OnMessageCallbackArgs] for an example on how to deserialize the message.
-    pub message: Vec<u8>,
-}
-/// Handler initialized by the canister
-/// and triggered by the CDK once an IC WebSocket message is received.
-type OnMessageCallback = fn(OnMessageCallbackArgs);
-
-/// Arguments passed to the `on_close` handler.
-pub struct OnCloseCallbackArgs {
-    pub client_principal: ClientPrincipal,
-}
-/// Handler initialized by the canister
-/// and triggered by the CDK once the WS Gateway closes the IC WebSocket connection
-/// for that client.
-type OnCloseCallback = fn(OnCloseCallbackArgs);
-
-/// Handlers initialized by the canister and triggered by the CDK.
-#[derive(Clone, Debug, Default, PartialEq)]
-pub struct WsHandlers {
-    pub on_open: Option<OnOpenCallback>,
-    pub on_message: Option<OnMessageCallback>,
-    pub on_close: Option<OnCloseCallback>,
-}
-
-impl WsHandlers {
-    fn call_on_open(&self, args: OnOpenCallbackArgs) {
-        if let Some(on_open) = self.on_open {
-            let res = panic::catch_unwind(|| {
-                on_open(args);
-            });
-
-            if let Err(e) = res {
-                custom_print!("Error calling on_open handler: {:?}", e);
-            }
-        }
-    }
-
-    fn call_on_message(&self, args: OnMessageCallbackArgs) {
-        if let Some(on_message) = self.on_message {
-            let res = panic::catch_unwind(|| {
-                on_message(args);
-            });
-
-            if let Err(e) = res {
-                custom_print!("Error calling on_message handler: {:?}", e);
-            }
-        }
-    }
-
-    fn call_on_close(&self, args: OnCloseCallbackArgs) {
-        if let Some(on_close) = self.on_close {
-            let res = panic::catch_unwind(|| {
-                on_close(args);
-            });
-
-            if let Err(e) = res {
-                custom_print!("Error calling on_close handler: {:?}", e);
-            }
-        }
-    }
-}
-
-/// Parameters for the IC WebSocket CDK initialization. For default parameters and simpler initialization, use [`WsInitParams::new`].
-#[derive(Clone)]
-pub struct WsInitParams {
-    /// The callback handlers for the WebSocket.
-    pub handlers: WsHandlers,
-    /// The maximum number of messages to be returned in a polling iteration.
-    /// Defaults to `10`.
-    pub max_number_of_returned_messages: usize,
-    /// The interval at which to send an acknowledgement message to the client,
-    /// so that the client knows that all the messages it sent have been received by the canister (in milliseconds).
-    ///
-    /// Must be greater than `keep_alive_timeout_ms`.
-    ///
-    /// Defaults to `60_000` (60 seconds).
-    pub send_ack_interval_ms: u64,
-    /// The delay to wait for the client to send a keep alive after receiving an acknowledgement (in milliseconds).
-    ///
-    /// Must be lower than `send_ack_interval_ms`.
-    ///
-    /// Defaults to `10_000` (10 seconds).
-    pub keep_alive_timeout_ms: u64,
-}
-
-impl WsInitParams {
-    /// Creates a new instance of WsInitParams, with default interval values.
-    pub fn new(handlers: WsHandlers) -> Self {
-        Self {
-            handlers,
-            ..Default::default()
-        }
-    }
-
-    fn get_handlers(&self) -> WsHandlers {
-        self.handlers.clone()
-    }
-
-    /// Checks the validity of the timer parameters.
-    /// `send_ack_interval_ms` must be greater than `keep_alive_timeout_ms`.
-    ///
-    /// # Traps
-    /// If `send_ack_interval_ms` <= `keep_alive_timeout_ms`.
-    fn check_validity(&self) {
-        if self.keep_alive_timeout_ms >= self.send_ack_interval_ms {
-            custom_trap!("send_ack_interval_ms must be greater than keep_alive_timeout_ms");
-        }
-    }
-
-    pub fn with_max_number_of_returned_messages(
-        mut self,
-        max_number_of_returned_messages: usize,
-    ) -> Self {
-        self.max_number_of_returned_messages = max_number_of_returned_messages;
-        self
-    }
-
-    pub fn with_send_ack_interval_ms(mut self, send_ack_interval_ms: u64) -> Self {
-        self.send_ack_interval_ms = send_ack_interval_ms;
-        self.check_validity();
-        self
-    }
-
-    pub fn with_keep_alive_timeout_ms(mut self, keep_alive_timeout_ms: u64) -> Self {
-        self.keep_alive_timeout_ms = keep_alive_timeout_ms;
-        self.check_validity();
-        self
-    }
-}
-
-impl Default for WsInitParams {
-    fn default() -> Self {
-        Self {
-            handlers: WsHandlers::default(),
-            max_number_of_returned_messages: DEFAULT_MAX_NUMBER_OF_RETURNED_MESSAGES,
-            send_ack_interval_ms: DEFAULT_SEND_ACK_INTERVAL_MS,
-            keep_alive_timeout_ms: DEFAULT_CLIENT_KEEP_ALIVE_TIMEOUT_MS,
-        }
-    }
-}
 
 /// Initialize the CDK by setting the callback handlers and the **principals** of the WS Gateways that
 /// will be polling the canister.
@@ -1231,4 +239,13 @@ pub fn ws_get_messages(args: CanisterWsGetMessagesArguments) -> CanisterWsGetMes
 pub fn ws_send(client_principal: ClientPrincipal, msg_bytes: Vec<u8>) -> CanisterWsSendResult {
     let client_key = get_client_key_from_principal(&client_principal)?;
     _ws_send(&client_key, msg_bytes, false)
+}
+
+/// Resets the internal state of the IC WebSocket CDK.
+///
+/// **Note:** You should only call this function in tests.
+pub fn wipe() {
+    reset_internal_state();
+
+    custom_print!("Internal state has been wiped!");
 }
