+++ src/ic-websocket-cdk/src/tests/unit_tests/mod.rs
@@ -1,3 +1,5 @@
+use std::{cell::RefCell, panic};
+
 use super::common;
 use crate::*;
 use proptest::prelude::*;
